pyproject.toml
[project]
name = "todo-pancakes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litestar[standard]>=2.17.0",
    "psycopg>=3.2.10",
    "sqlalchemy>=2.0.43",
    "strawberry-graphql[debug-server]>=0.282.0",
]
